[C8] Split featuregraph.go by responsibility (#2228)
model/exchange/translators/inventor/ipt/featuregraph.go was 536 LOC, over
the 500-line ceiling. Split the .ipt feature-graph reader along its three
responsibilities; each file is now under the ceiling:

- featuregraph.go (177): graph decoding — the raw DC node list (dcNodes),
  the pre-2023 content-header normalisation, reference resolution (nodeAt)
  and the property ref-list (featureProperties).
- featuregraph_payload.go (225): the per-node payload readers — Parameter
  f64, Boolean byte, enum u16, DirectionAxis vector, and the feature-relative
  readers (featureParameter/featureBoolean/featureEnum, extrudeDirection/
  Depth/Operation/ThroughAll).
- featuregraph_deps.go (167): dependency resolution — the exported queries
  that walk the graph to bind each extrude to its depth and profile sketch
  and the part's revolve profile (ExtrudeDepths/ExtrudeProfiles/
  RevolveProfileSketch + the patch->faceBound->sketch resolution).

Behaviour-preserving: every const and function body is byte-identical, only
moved; consts grouped with their primary readers. The ipt corpus passes
before and after (GOWORK=off, this is a separate module) with no golden
updates; module lint issue count unchanged (32, all pre-existing in other
files). Net delta of recognizers / tolerance constants / fallback sites /
type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2228.

(cherry picked from commit 1a460a52e1f1365321366c607fabe0f14f83199b)
