+++ model/exchange/translators/inventor/ipt/featuregraph.go
@@ -12,10 +12,14 @@ package ipt
 // (_iptformat/reference-InventorLoader): Read_90874D91 "Feature", Read_90874D26 "Parameter", and
 // Create_FxExtrude_New's property indices. Cross-checked against files: 14_box_two's two extrudes
 // resolve to Parameter nodes holding exactly 1 cm and 2 cm.
+//
+// This file is the GRAPH DECODING: the raw node list (dcNodes), the pre-2023 content-header
+// normalisation, reference resolution (nodeAt) and the property ref-list (featureProperties). The
+// per-node value readers live in featuregraph_payload.go; the profile/axis dependency queries in
+// featuregraph_deps.go.
 
 import (
 	"encoding/binary"
-	"math"
 )
 
 const (
@@ -23,29 +27,6 @@ const (
 	parameterNodeType = 0x90874D26 // Read_90874D26 — a named parameter with its value
 	directionAxisType = 0xCE52DF40 // an extrude's direction; absent on revolve/other features
 )
-
-// Property indices on a feature node, from InventorLoader's Create_FxExtrude_New. A feature whose
-// property 2 is a DirectionAxis is an extrude, and its property 4 is the Parameter holding the
-// depth (dimLength1).
-const (
-	propReversed  = 3 // Boolean: grow against the profile normal
-	propDirection = 2
-	propDistance  = 4
-	propMidplane  = 7    // Boolean: grow half each way about the sketch plane
-	propDistance2 = 0x1B // Parameter: the second direction's length (a two-sided extrude)
-)
-
-// booleanNodeType is InventorLoader's Read_90874D28 "Boolean".
-const booleanNodeType = 0x90874D28
-
-// directionVecOffset is where a DirectionAxis holds its unit direction. InventorLoader
-// Read_CE52DF40: header, an f64, the >2017 4-byte gap, then the f64x3 'dir'. Self-validating —
-// every one of the corpus's 252 extrude directions reads back with |d| = 1 here.
-const directionVecOffset = 50
-
-// booleanNameOffset is where a Boolean node's length-prefixed UTF-16 name begins: the content
-// header runs to 34, exactly as on a Parameter.
-const booleanNameOffset = 34
 
 // Byte layout of a feature node's payload, for this Inventor generation (block size 0, and the
 // post-2023 4-byte gap in the content header). The property list is a List2 of cross-references:
@@ -193,344 +174,4 @@ func featureProperties(pay []byte) ([]int, bool) {
 		out[i] = int(binary.LittleEndian.Uint32(pay[ds+i*4:]) & refIndexMask)
 	}
 	return out, true
-}
-
-// parameterValue reads a Parameter node's nominal value. Layout after the content header: a
-// length-prefixed UTF-16 name at 34, a 4-byte gap, the unit and formula refs, then the f64.
-func parameterValue(pay []byte) (float64, bool) {
-	const nameLenOffset = 34
-	if nameLenOffset+4 > len(pay) {
-		return 0, false
-	}
-	n := int(binary.LittleEndian.Uint32(pay[nameLenOffset:]))
-	if n < 0 || n > 256 {
-		return 0, false
-	}
-	at := nameLenOffset + 4 + 2*n + 12 // name, gap(4), unit ref(4), formula ref(4)
-	if at+8 > len(pay) {
-		return 0, false
-	}
-	v := math.Float64frombits(binary.LittleEndian.Uint64(pay[at:]))
-	if math.IsNaN(v) || math.IsInf(v, 0) {
-		return 0, false
-	}
-	return v, true
-}
-
-// booleanValue reads a Boolean node's value. Layout after the content header (InventorLoader
-// Read_90874D28): a length-prefixed UTF-16 name, a 4-byte gap, then ONE byte — NOT the u16 an enum
-// node carries at 36. Reading it at the enum offset lands on the name-length field, which is 0 on
-// every node here, so the whole corpus silently reported false.
-func booleanValue(pay []byte) (bool, bool) {
-	if booleanNameOffset+4 > len(pay) {
-		return false, false
-	}
-	n := int(binary.LittleEndian.Uint32(pay[booleanNameOffset:]))
-	if n < 0 || n > 256 {
-		return false, false
-	}
-	at := booleanNameOffset + 4 + 2*n + 4 // name, gap(4)
-	if at >= len(pay) {
-		return false, false
-	}
-	return pay[at] != 0, true
-}
-
-// featureBoolean reads the Boolean a feature's property i references (false when absent).
-func featureBoolean(nodes []dcNode, props []int, i int) bool {
-	if i >= len(props) {
-		return false
-	}
-	n, ok := nodeAt(nodes, props[i])
-	if !ok || n.typ != booleanNodeType {
-		return false
-	}
-	v, ok := booleanValue(n.payload)
-	return ok && v
-}
-
-// enumValueOf reads a node's 16-bit enum value (an operation, extent, or hole-kind enum), and whether
-// the payload is long enough to hold one.
-func enumValueOf(n dcNode) (uint16, bool) {
-	if len(n.payload) < enumValueOffset+2 {
-		return 0, false
-	}
-	return binary.LittleEndian.Uint16(n.payload[enumValueOffset:]), true
-}
-
-// featureEnum reads the enum value the feature's i-th property names, and whether that property
-// resolves to a node carrying one — the generic form of extrudeExtentEnum for any property slot
-// (a hole's extent sits at a different index than an extrude's).
-func featureEnum(nodes []dcNode, props []int, i int) (uint16, bool) {
-	if i >= len(props) {
-		return 0, false
-	}
-	n, ok := nodeAt(nodes, props[i])
-	if !ok {
-		return 0, false
-	}
-	return enumValueOf(n)
-}
-
-// extrudeDirection reads the unit direction the feature's DirectionAxis names, and whether it read
-// at all. This is the vector the extrude actually grows along; `reversed` flips THIS, not the
-// profile normal — so reversed cannot be interpreted without it (see Extrude.Dir).
-func extrudeDirection(nodes []dcNode, props []int) ([3]float64, bool) {
-	var d [3]float64
-	if propDirection >= len(props) {
-		return d, false
-	}
-	n, ok := nodeAt(nodes, props[propDirection])
-	if !ok || n.typ != directionAxisType || len(n.payload) < directionVecOffset+24 {
-		return d, false
-	}
-	for i := 0; i < 3; i++ {
-		d[i] = math.Float64frombits(binary.LittleEndian.Uint64(n.payload[directionVecOffset+8*i:]))
-	}
-	l := math.Sqrt(d[0]*d[0] + d[1]*d[1] + d[2]*d[2])
-	if math.IsNaN(l) || math.Abs(l-1) > 1e-6 {
-		return [3]float64{}, false // not a unit vector ⇒ not this layout; decline rather than guess
-	}
-	return d, true
-}
-
-// featureParameter reads the Parameter value a feature's property i references (0 when absent).
-func featureParameter(nodes []dcNode, props []int, i int) float64 {
-	if i >= len(props) {
-		return 0
-	}
-	n, ok := nodeAt(nodes, props[i])
-	if !ok || n.typ != parameterNodeType {
-		return 0
-	}
-	v, ok := parameterValue(n.payload)
-	if !ok {
-		return 0
-	}
-	return v
-}
-
-// extrudeDepth returns the depth a feature node references, and whether that node is an extrude at
-// all: an extrude is the feature whose direction property is a DirectionAxis (a revolve, hole or
-// fillet has something else there), and its distance property must resolve to a Parameter.
-func extrudeDepth(nodes []dcNode, pay []byte) (float64, bool) {
-	props, ok := featureProperties(pay)
-	if !ok || len(props) <= propDistance {
-		return 0, false
-	}
-	dir, ok := nodeAt(nodes, props[propDirection])
-	if !ok || dir.typ != directionAxisType {
-		return 0, false
-	}
-	dist, ok := nodeAt(nodes, props[propDistance])
-	if !ok || dist.typ != parameterNodeType {
-		return 0, false
-	}
-	return parameterValue(dist.payload)
-}
-
-// ExtrudeDepths returns the depth of every extrude feature, in feature order, each read from that
-// feature's own referenced distance parameter.
-func ExtrudeDepths(d *Document) []float64 {
-	nodes := dcNodes(d)
-	var out []float64
-	for _, n := range nodes {
-		if n.typ != featureNodeType {
-			continue
-		}
-		if v, ok := extrudeDepth(nodes, n.payload); ok {
-			out = append(out, v)
-		}
-	}
-	return out
-}
-
-// The profile side of the feature graph: an extrude names a BoundaryPatch (property 1), and a
-// FaceBound node points BACK at that patch while naming the Sketch2D it bounds. So the profile
-// binding is patch -> faceBound -> sketch, which is what replaces "extrude i consumes sketch i" —
-// a guess that is provably wrong on real parts (BigChunkyPlate's 13th extrude uses its 15th
-// sketch). Grounded in InventorLoader: Read_22947391 "BoundaryPatch", Read_424EB7D7 "FaceBound",
-// Read_F9884C43 "FaceBoundOuter", ReadHeaderFaceBound.
-const (
-	boundaryPatchNodeType  = 0x22947391
-	faceBoundNodeType      = 0x424EB7D7
-	faceBoundOuterNodeType = 0xF9884C43
-)
-
-// propProfile is the extrude's BoundaryPatch property (InventorLoader Create_FxExtrude_New's
-// `patch = getProperty(properties, 0x01)`).
-const propProfile = 1
-
-// faceBoundHeaderEnd is where a FaceBound's three reference lists begin; its `sketch` and `proxy`
-// refs follow them and a u32 pair.
-const faceBoundHeaderEnd = 30
-
-// ExtrudeProfiles returns, for each extrude in feature order, the index of the sketch it actually
-// consumes (indices into GraphSketches). An extrude whose profile can't be resolved yields -1, so
-// callers skip it rather than build on a guessed sketch.
-func ExtrudeProfiles(d *Document) []int {
-	nodes := dcNodes(d)
-	_, sketchIndex := sketchOrdinals(nodes)
-	patchToSketch := boundPatchSketches(nodes, sketchIndex)
-	var out []int
-	for _, n := range nodes {
-		if n.typ != featureNodeType {
-			continue
-		}
-		if _, ok := extrudeDepth(nodes, n.payload); !ok {
-			continue // not an extrude; keep this list aligned with DecodeExtrudes
-		}
-		out = append(out, profileOf(n.payload, patchToSketch))
-	}
-	return out
-}
-
-// profileOf maps one extrude to its sketch index via its BoundaryPatch property.
-func profileOf(pay []byte, patchToSketch map[int]int) int {
-	props, ok := featureProperties(pay)
-	if !ok || len(props) <= propProfile {
-		return -1
-	}
-	if i, ok := patchToSketch[props[propProfile]]; ok {
-		return i
-	}
-	return -1
-}
-
-// revolveAxisNodeType is the node a Revolution feature names at property 2 (propDirection), in
-// place of an extrude's DirectionAxis: its own axis/centreline reference. It is what tells a revolve
-// feature apart from an extrude (which names 0xCE52DF40 there) — verified on the ReelToReel revolve
-// parts, where every revolve's property 2 resolves to this type and no extrude's does.
-const revolveAxisNodeType = 0x8EF06C89
-
-// RevolveProfileSketch returns the GraphSketches index of the sketch the part's Revolution feature
-// consumes as its profile, resolved through the feature's BoundaryPatch property — the SAME
-// patch -> FaceBound -> Sketch2D chain ExtrudeProfiles uses. This replaces GUESSING the revolve base
-// (the first sketch that merely looks revolvable), which mis-picks a CUT profile on a multi-feature
-// machined part and revolves garbage. A revolve feature is a generic Fx feature that is not an
-// extrude (no depth) and names its axis at property 2 as a revolveAxisNodeType node. ok=false when
-// the part has no such feature or its profile patch can't be mapped, so callers keep their fallback.
-func RevolveProfileSketch(d *Document) (int, bool) {
-	nodes := dcNodes(d)
-	_, sketchIndex := sketchOrdinals(nodes)
-	patchToSketch := boundPatchSketches(nodes, sketchIndex)
-	for _, n := range nodes {
-		if n.typ != featureNodeType {
-			continue
-		}
-		if _, isExtrude := extrudeDepth(nodes, n.payload); isExtrude {
-			continue
-		}
-		props, ok := featureProperties(n.payload)
-		if !ok || len(props) <= propProfile || !isRevolveFeature(nodes, props) {
-			continue
-		}
-		if i, ok := patchToSketch[props[propProfile]]; ok {
-			return i, true
-		}
-	}
-	return 0, false
-}
-
-// isRevolveFeature reports whether a feature's properties are a revolve's — property 2 (propDirection)
-// names a revolveAxisNodeType node, where an extrude names a DirectionAxis and a hole names its kind
-// at property 0.
-func isRevolveFeature(nodes []dcNode, props []int) bool {
-	if len(props) <= propDirection {
-		return false
-	}
-	ax, ok := nodeAt(nodes, props[propDirection])
-	return ok && ax.typ == revolveAxisNodeType
-}
-
-// boundPatchSketches maps each BoundaryPatch ordinal to the index of the sketch that bounds it,
-// by walking the FaceBound nodes that name the patch as their proxy.
-func boundPatchSketches(nodes []dcNode, sketchIndex map[int]int) map[int]int {
-	out := map[int]int{}
-	for _, n := range nodes {
-		if n.typ != faceBoundNodeType && n.typ != faceBoundOuterNodeType {
-			continue
-		}
-		sketchRef, patchRef, ok := faceBoundRefs(n.payload)
-		if !ok {
-			continue
-		}
-		if i, ok := sketchIndex[sketchRef]; ok {
-			out[patchRef] = i
-		}
-	}
-	return out
-}
-
-// faceBoundRefs reads a FaceBound's `sketch` and `proxy` references, which follow its three
-// variable-length reference lists and a u32 pair.
-func faceBoundRefs(pay []byte) (sketchRef, patchRef int, ok bool) {
-	i := faceBoundHeaderEnd
-	for k := 0; k < 3; k++ {
-		_, next, ok := refList2(pay, i)
-		if !ok {
-			return 0, 0, false
-		}
-		i = next
-	}
-	i += 8 // the u32 pair before the refs
-	if i+8 > len(pay) {
-		return 0, 0, false
-	}
-	sketchRef = int(binary.LittleEndian.Uint32(pay[i:]) & refIndexMask)
-	patchRef = int(binary.LittleEndian.Uint32(pay[i+4:]) & refIndexMask)
-	return sketchRef, patchRef, true
-}
-
-// partFeatureOperationEnumType is the enum node an extrude names as its FIRST property: the boolean
-// operation it applies (InventorLoader Read_729ABE28 "PartFeatureOperationEnum", whose values —
-// 1 NewBody, 2 Cut, 3 Join, 4 Intersection — are the OpNewBody..OpIntersect constants).
-const partFeatureOperationEnumType = 0x729ABE28
-
-// propOperation is the extrude's operation property (InventorLoader Create_FxExtrude reads
-// `getPropertyValue(properties, 0x00, 'value')`).
-const propOperation = 0
-
-// enumValueOffset is where an enum node holds its value: the content header runs to 34, then a
-// 2-byte type discriminator.
-const enumValueOffset = 36
-
-// extrudeOperation returns the boolean operation a feature node NAMES, and whether it names one.
-// The predecessor took the i-th enum node found by scanning the segment's bytes, which pairs
-// operations with extrudes by position — the same guess that made every extrude take the wrong
-// depth, and just as wrong on a part whose features are not authored in scan order.
-func extrudeOperation(nodes []dcNode, pay []byte) (int, bool) {
-	props, ok := featureProperties(pay)
-	if !ok || len(props) <= propOperation {
-		return 0, false
-	}
-	n, ok := nodeAt(nodes, props[propOperation])
-	if !ok || n.typ != partFeatureOperationEnumType || len(n.payload) < enumValueOffset+2 {
-		return 0, false
-	}
-	return int(binary.LittleEndian.Uint16(n.payload[enumValueOffset:])), true
-}
-
-// propExtent is the extrude's termination property (InventorLoader Create_FxExtrude_New reads
-// `extend = getPropertyValue(properties, 0x06, 'value')`).
-const propExtent = 6
-
-// extentAll is the `extend` value meaning the extrude runs through ALL the material, taking its
-// length from the body rather than from a parameter. Verified on BigChunkyPlate: its four
-// distance-less extrudes read 5 here while every measured one reads 1.
-const extentAll = 5
-
-// extentEnumNodeType is the PartFeatureExtentEnum node (InventorLoader Read_92637D29).
-const extentEnumNodeType = 0x92637D29
-
-// extrudeThroughAll reports whether a feature terminates by running through all the material. Such
-// an extrude has NO distance: its depth parameter decodes as 0, and building that as a length makes
-// a zero-thickness body.
-func extrudeThroughAll(nodes []dcNode, pay []byte) bool {
-	props, ok := featureProperties(pay)
-	if !ok {
-		return false
-	}
-	v, ok := extrudeExtentEnum(nodes, props)
-	return ok && v == extentAll
 }
